+++ R/create_model_tasks.R
@@ -80,7 +80,7 @@
 #'       ),
 #'       create_output_type_median(
 #'         is_required = FALSE,
-#'         value_type = "numeric"
+#'         value_type = "double"
 #'       ),
 #'       create_output_type_quantile(
 #'         required = c(0.25, 0.5, 0.75),
@@ -137,7 +137,7 @@
 #'       )
 #'     ),
 #'     output_type = create_output_type(
-#'       create_output_type_categorical(
+#'       create_output_type_pmf(
 #'         required = c(
 #'           "large_decrease",
 #'           "decrease",
@@ -146,7 +146,7 @@
 #'           "large_increase"
 #'         ),
 #'         optional = NULL,
-#'         value_type = "numeric"
+#'         value_type = "double"
 #'       )
 #'     ),
 #'     target_metadata = create_target_metadata(
@@ -155,7 +155,7 @@
 #'         target_name = "Weekly influenza hospitalization rate change",
 #'         target_units = "rate per 100,000 population",
 #'         target_keys = list(target = "flu hosp rt chng"),
-#'         target_type = "categorical",
+#'         target_type = "nominal",
 #'         is_step_ahead = TRUE,
 #'         time_unit = "week"
 #'       )